pyproject.toml
[project]
name = "kkgepo"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "edoelas", email = "miego.cvcv@gmail.com" }
]
dependencies = [
    "iterfzf>=1.4.0.54.3",
    "kr8s>=0.18.1",
    "typer>=0.13.1",
    "ipython>=8.29.0",
    "pyyaml>=6.0.2",
    "jsonpath-ng>=1.7.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kkgepo"]
